requirements.txt
numpy>=1.21.0
open3d>=0.15.0
PyYAML>=6.0
matplotlib>=3.5.0